+++ ChangeLog.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
-- `remove`, `distinct`, and `sorted` to list of transducer functions
+- `remove`, `distinct`, `reverse`, and `sorted` to list of transducer functions
 
 ### Fixed